+++ CHANGES.md
@@ -1,4 +1,6 @@
 #### [unreleased]
+
+#### 11.1.7 / 2022-10-24
 * load `wp-admin/includes/file.php` for when move_dir() in core, avoid redeclaration error
 * allow updating of Git Remote Updater
 * now using `str_contains`